Auto applmk2/requirements.txt
selenium==4.23.1
jinja2==3.1.4
PyYAML==6.0.2
pandas==2.2.2
python-dotenv==1.0.1
webdriver-manager==4.0.1
